requirements.txt
numpy
matplotlib
Orange3
scikit-bio
biopython